+++ src/Interfaces/AnimationFactoryInterface.php
@@ -7,17 +7,13 @@ namespace Intervention\Image\Interfaces;
 interface AnimationFactoryInterface
 {
     /**
-     * Create the end product of the factory statically by calling given callable
-     */
-    public static function build(DriverInterface $driver, callable $animation): ImageInterface;
-
-    /**
-     * Resolve image from given source and add it as new animation frame with specific delay.
+     * Resolve image from given source and add it as new animation
+     * frame with specific delay in seconds.
      */
     public function add(mixed $source, float $delay = 1): self;
 
     /**
-     * Create image instance.
+     * Build ready-made animation as end product.
      */
     public function animation(): ImageInterface;
 }