+++ tests/Unit/Drivers/Imagick/DriverTest.php
@@ -27,7 +27,6 @@ use Intervention\Image\Interfaces\SpecializableInterface;
 use Intervention\Image\MediaType;
 use Intervention\Image\Tests\BaseTestCase;
 use Intervention\Image\Tests\Providers\InputDataProvider;
-use Intervention\Image\Tests\Resource;
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\DataProviderExternal;
@@ -55,20 +54,6 @@ final class DriverTest extends BaseTestCase
         $this->assertInstanceOf(ImageInterface::class, $image);
         $this->assertEquals(3, $image->width());
         $this->assertEquals(2, $image->height());
-    }
-
-    public function testCreateAnimation(): void
-    {
-        $image = $this->driver->createAnimation(function ($animation): void {
-            $animation->add(Resource::create('red.gif')->path(), .25);
-            $animation->add(Resource::create('green.gif')->path(), .25);
-        })->setLoops(5);
-        $this->assertInstanceOf(ImageInterface::class, $image);
-
-        $this->assertEquals(16, $image->width());
-        $this->assertEquals(16, $image->height());
-        $this->assertEquals(5, $image->loops());
-        $this->assertEquals(2, $image->count());
     }
 
     /**